backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
mysqlclient
django-cors-headers
pandas
django-redis
redis
razorpay
python-dotenv
django-anymail
